consumer group testing
