+++ packages/message-db-consumer/tests/source.test.mts
@@ -145,53 +145,3 @@ test("it fails fast", async () => {
 
   await expect(src.start(ctrl.signal)).rejects.toThrow("failed")
 })
-
-describe("MessageDbCategoryReader", () => {
-  const connectionString =
-    process.env.MDB_CONN_STR ?? "postgres://message_store:@localhost:5432/message_store"
-  const pool = new Pool({ connectionString })
-  const conn = MessageDbConnection.create(pool)
-  const category = randomUUID().replace(/-/g, "")
-  const streamId = randomUUID().replace(/-/g, "")
-  const streamName = StreamName.compose(category, streamId)
-  beforeAll(async () => {
-    await conn.write.writeMessages(streamName, Array(100).fill({ type: "TestEvent" }), -1n)
-    await conn.write.writeSingleMessage(streamName, { type: "ExclusiveEvent" }, null)
-  })
-  test("Can read a category", async () => {
-    const reader = new MessageDbCategoryReader(pool)
-    const batch = await reader.readCategoryMessages({
-      category,
-      fromPositionInclusive: 0n,
-      batchSize: 10,
-    })
-    expect(batch.messages).toHaveLength(10)
-    expect(batch.isTail).toBe(false)
-    expect(batch.checkpoint).toBeGreaterThan(0n)
-  })
-  test("Can not read a category with a condition if sql_condition is off", async () => {
-    const reader = new MessageDbCategoryReader(pool)
-    const batch = reader.readCategoryMessages({
-      category,
-      fromPositionInclusive: 0n,
-      batchSize: 10,
-      condition: "messages.type = 'ExclusiveEvent'",
-    })
-
-    await expect(batch).rejects.toThrow("SQL condition is not activated")
-  })
-
-  test("Can read a category with a condition", async () => {
-    await pool.query("SELECT set_config('message_store.sql_condition', 'on', true)", [])
-    const reader = new MessageDbCategoryReader(pool)
-    const batch = await reader.readCategoryMessages({
-      category,
-      fromPositionInclusive: 0n,
-      batchSize: 10,
-      condition: "messages.type = 'ExclusiveEvent'",
-    })
-    expect(batch.messages).toHaveLength(1)
-    expect(batch.isTail).toBe(true)
-    expect(batch.checkpoint).toBeGreaterThan(0n)
-  })
-})
